fixed compilation with new 10.5 compatible headers

+++ Foundation/Sources/NSCoder.m
@@ -23,7 +23,7 @@
 
 - (unsigned int) systemVersion					{ return 104; }
 
-- (unsigned int) versionForClassName:(NSString*)className 	
+- (NSInteger) versionForClassName:(NSString*)className 	
 { 
 	SUBCLASS return NSNotFound; 
 }
@@ -169,7 +169,7 @@
 - (void) encodeObject:(id) object forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeInt:(int) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeInt64:(long long) value forKey:(NSString *) key; { SUBCLASS }
-- (void) encodeInt32:(long) value forKey:(NSString *) key; { SUBCLASS }
+- (void) encodeInt32:(int32_t) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeFloat:(float) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeDouble:(double) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeConditionalObject:(id) object forKey:(NSString *) key; { SUBCLASS }
@@ -184,7 +184,7 @@
 - (id) decodeObjectForKey:(NSString *) key; { return SUBCLASS; }
 - (int) decodeIntForKey:(NSString *) key; { SUBCLASS; return 0; }
 - (long long) decodeInt64ForKey:(NSString *) key; { SUBCLASS; return 0; }
-- (long) decodeInt32ForKey:(NSString *) key; { SUBCLASS; return 0; }
+- (int32_t) decodeInt32ForKey:(NSString *) key; { SUBCLASS; return 0; }
 - (float) decodeFloatForKey:(NSString *) key; { SUBCLASS; return 0.0; }
 - (double) decodeDoubleForKey:(NSString *) key; { SUBCLASS; return 0.0; }
 - (const unsigned char *) decodeBytesForKey:(NSString *) key returnedLength:(unsigned *) num; { SUBCLASS; return NULL; }
